# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Career Guidance System API
 * OpenAPI spec version: 0.1.0
 */
export interface ResumeFileInput {
  file: Blob;
}

export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  targetRole?: string | null;
  /** @nullable */
  currentRole?: string | null;
  skills?: string[];
  resumeCount?: number;
  predictionCount?: number;
  createdAt: string;
}

export interface ProfileUpdate {
  name?: string;
  bio?: string;
  targetRole?: string;
  currentRole?: string;
}

export interface Resume {
  id: number;
  userId: number;
  filename: string;
  /** @nullable */
  score: number | null;
  skills: string[];
  missingSkills?: string[];
  /** @nullable */
  rawText?: string | null;
  createdAt: string;
}

export interface ResumeParseResult {
  resume: Resume;
  skills: string[];
  missingSkills: string[];
  score: number;
}

export interface ScoreInput {
  resumeId: number;
  /** @nullable */
  targetRole?: string | null;
}

export interface PredictionResult {
  resumeScore: number;
  placementEligibility: boolean;
  placementProbability?: number;
  careerDomain: string;
  clusterGroup: string;
  performancePrediction?: number;
  skillGaps?: string[];
}

export interface Prediction {
  id: number;
  userId: number;
  resumeId: number;
  resumeScore: number;
  placementEligibility: boolean;
  /** @nullable */
  placementProbability?: number | null;
  careerDomain: string;
  /** @nullable */
  clusterGroup?: string | null;
  /** @nullable */
  performancePrediction?: number | null;
  skillGaps?: string[];
  createdAt: string;
}

export interface SuggestionInput {
  resumeId: number;
  /** @nullable */
  targetRole?: string | null;
}

export interface Course {
  title: string;
  provider: string;
  url: string;
  level: string;
}

export interface CareerSuggestionResult {
  careerGuidance: string;
  skillImprovements: string[];
  courses: Course[];
  interviewTips: string[];
  resumeTips: string[];
}

export interface RoadmapInput {
  skills: string[];
  targetRole: string;
}

export interface RoadmapPhase {
  phase: number;
  title: string;
  skills: string[];
  duration: string;
  resources?: string[];
}

export interface RoadmapResult {
  targetRole: string;
  phases: RoadmapPhase[];
  estimatedMonths: number;
}

export interface Suggestion {
  id: number;
  userId: number;
  type: string;
  content: string;
  createdAt: string;
}

export interface DomainStat {
  domain: string;
  count: number;
  confidence: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  createdAt: string;
}

export interface DashboardStats {
  totalResumes: number;
  averageScore: number;
  bestScore: number;
  placementEligible: boolean;
  placementProbability?: number;
  topSkills: string[];
  missingSkills?: string[];
  careerDomains: DomainStat[];
  recentActivity: ActivityItem[];
}

export interface SkillTrend {
  skill: string;
  count: number;
  category: string;
}

export interface ScoreHistoryPoint {
  date: string;
  score: number;
  resumeId: number;
}