# lib/api-zod/src/generated/types/prediction.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Career Guidance System API
 * OpenAPI spec version: 0.1.0
 */

export interface Prediction {
  id: number;
  userId: number;
  resumeId: number;
  resumeScore: number;
  placementEligibility: boolean;
  /** @nullable */
  placementProbability?: number | null;
  careerDomain: string;
  /** @nullable */
  clusterGroup?: string | null;
  /** @nullable */
  performancePrediction?: number | null;
  skillGaps?: string[];
  createdAt: Date;
}